+++ packages/vsce/src/webviews/common/messages.ts
@@ -31,5 +31,5 @@ export type ExtensionToWebviewMessage = {
   resourceIconPath: IResourceInspectorIconPath;
   humanReadableNamePlural: string;
   humanReadableNameSingular: string;
-  hasDatasetCommand: boolean;
+  shouldRenderDatasetLinks: boolean;
 };