+++ packages/vsce/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to the "cics-extension-for-zowe" extension will be documented in this file.
 
 ## Recent Changes
-- Enhancement: Enabled CICS attribute links to Data Sets in Zowe Explorer. [#590] (https://github.com/zowe/cics-for-zowe-client/issues/590)
+- Enhancement: Enabled CICS attribute links to Data Sets in Zowe Explorer. [#590](https://github.com/zowe/cics-for-zowe-client/issues/590)
 
 ## `3.19.0`